requirements.txt
streamlit
langchain
chromadb
sentence-transformers
groq
yfinance
requests
python-dotenv
openai